+++ packages/stargate/src/fee.ts
@@ -8,6 +8,18 @@ import { coins } from "@cosmjs/proto-signing";
 export type FeeTable = Record<string, StdFee>;
 
 /**
+ * Denom checker for the Cosmos SDK 0.42 denom pattern
+ * (https://github.com/cosmos/cosmos-sdk/blob/v0.42.4/types/coin.go#L599-L601).
+ *
+ * This is like a regexp but with helpful error messages.
+ */
+function checkDenom(denom: string): void {
+  if (denom.length < 3 || denom.length > 128) {
+    throw new Error("Denom must be between 3 and 128 characters");
+  }
+}
+
+/**
  * A gas price, i.e. the price of a single gas. This is typically a fraction of
  * the smallest fee token unit, such as 0.012utoken.
  *
@@ -24,16 +36,20 @@ export class GasPrice {
 
   /**
    * Parses a gas price formatted as `<amount><denom>`, e.g. `GasPrice.fromString("0.012utoken")`.
+   *
+   * The denom must match the Cosmos SDK 0.42 pattern (https://github.com/cosmos/cosmos-sdk/blob/v0.42.4/types/coin.go#L599-L601).
+   * See `GasPrice` in @cosmjs/stargate for a more generic matcher.
+   *
+   * Separators are not yet supported.
    */
   public static fromString(gasPrice: string): GasPrice {
-    const matchResult = gasPrice.match(/^([0-9.]+?)([a-z]+)$/);
+    // Use Decimal.fromUserInput and checkDenom for detailed checks and helpful error messages
+    const matchResult = gasPrice.match(/^([0-9.]+)([a-z][a-z0-9]*)$/i);
     if (!matchResult) {
       throw new Error("Invalid gas price string");
     }
     const [_, amount, denom] = matchResult;
-    if (denom.length < 3 || denom.length > 127) {
-      throw new Error("Gas price denomination must be between 3 and 127 characters");
-    }
+    checkDenom(denom);
     const fractionalDigits = 18;
     const decimalAmount = Decimal.fromUserInput(amount, fractionalDigits);
     return new GasPrice(decimalAmount, denom);